@@ -11,49 +11,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.10",
-    "@angular/cdk": "^7.3.5",
-    "@angular/common": "~7.2.5",
-    "@angular/compiler": "~7.2.5",
-    "@angular/core": "~7.2.5",
-    "@angular/forms": "~7.2.5",
-    "@angular/material": "^7.3.5",
-    "@angular/platform-browser": "~7.2.5",
-    "@angular/platform-browser-dynamic": "~7.2.5",
-    "@angular/router": "~7.2.5",
-    "@auth0/angular-jwt": "^2.1.0",
+    "@angular/animations": "^9.0.0-next.3",
+    "@angular/cdk": "^8.1.4",
+    "@angular/common": "~9.0.0-next.3",
+    "@angular/compiler": "~9.0.0-next.3",
+    "@angular/core": "~9.0.0-next.3",
+    "@angular/forms": "~9.0.0-next.3",
+    "@angular/material": "^8.1.4",
+    "@angular/material-moment-adapter": "^8.1.4",
+    "@angular/platform-browser": "~9.0.0-next.3",
+    "@angular/platform-browser-dynamic": "~9.0.0-next.3",
+    "@angular/router": "~9.0.0-next.3",
+    "@auth0/angular-jwt": "^3.0.0",
     "@ng-bootstrap/ng-bootstrap": "^4.0.3",
     "@swimlane/ngx-datatable": "^14.0.0",
-    "alertifyjs": "^1.11.2",
-    "core-js": "^2.5.4",
-    "ngx-bootstrap": "^3.2.0",
+    "alertifyjs": "^1.11.4",
+    "bootstrap": "^4.3.1",
+    "bootswatch": "^4.3.1",
+    "core-js": "^3.2.1",
+    "moment": "^2.24.0",
+    "ngx-bootstrap": "^5.1.1",
+    "ngx-gallery": "^5.10.0",
     "ngx-mask": "^7.9.10",
-    "ngx-spinner": "^7.0.0",
+    "ngx-spinner": "^7.2.0",
     "ngx-trend": "^3.4.1",
-    "rxjs": "~6.4.0",
+    "rxjs": "^6.5.2",
     "sweetalert2": "^8.2.3",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "~7.3.1",
-    "@angular/compiler-cli": "^7.2.5",
-    "@angular/language-service": "~7.2.5",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.4"
+    "@angular-devkit/build-angular": "~0.803.0",
+    "@angular/cli": "~8.3.0",
+    "@angular/compiler-cli": "^9.0.0-next.3",
+    "@angular/language-service": "~9.0.0-next.3",
+    "@types/jasmine": "^3.4.0",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "^8.9.5",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "^3.4.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "^5.4.2",
+    "ts-node": "^8.3.0",
+    "tslint": "^5.19.0",
+    "typescript": "^3.5.3"
   }
 }
