@@ -89,10 +89,11 @@ namespace Admin.API.Migrations
                         .IsRequired()
                         .HasMaxLength(30);
 
-                    b.Property<int>("UserId")
-                        .HasColumnName("CreoUserId");
+                    b.Property<int>("UserId");
 
                     b.HasKey("Id");
+
+                    b.HasIndex("UserId");
 
                     b.ToTable("Tickets");
                 });
@@ -246,6 +247,14 @@ namespace Admin.API.Migrations
                         .WithOne("Equipo")
                         .HasForeignKey("Admin.API.Models.Equipo", "UserId")
                         .OnDelete(DeleteBehavior.SetNull);
+                });
+
+            modelBuilder.Entity("Admin.API.Models.Ticket", b =>
+                {
+                    b.HasOne("Admin.API.Models.User")
+                        .WithMany("Tickets")
+                        .HasForeignKey("UserId")
+                        .OnDelete(DeleteBehavior.Cascade);
                 });
 
             modelBuilder.Entity("Admin.API.Models.UserRole", b =>
