@@ -8,21 +8,36 @@ namespace Admin.API.Models
     {
         public int Id { get; set; }
 
-        [Column("CreoUserId")]
+        // [Column("CreoUserId")]u
         public int UserId { get; set; }
 
         public DateTime FechaAlta { get; set; }
-        
+
         [Required]
         [MaxLength(30)]
         [StringLength(30, ErrorMessage = "El nombre de equipo es obligatorio mi rey")]
         public string Titulo { get; set; }
 
         public string Mensaje { get; set; }
-
+        // [Column(TypeName = "decimal(5, 2)")]
         public byte Prioridad { get; set; }
 
         public byte Status { get; set; }
-        
+
+        public User User { get; set; }
+
     }
+
+    public enum Prioridad
+    {
+        Normal = 1, Alta = 2, Urgente = 3
+    }
+
+    public enum Status
+    {
+        Abierto = 1, EnProceso = 2, Cerrado = 3, ReAbrir = 4
+    }
+
+    //    bitacoraImportacion.TipoProceso = (int)TipoProceso.Nuevo;
+
 }
