@@ -1,5 +1,5 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
+import { NgModule, LOCALE_ID } from '@angular/core';
 import { JwtModule } from '@auth0/angular-jwt';
 
 // Rutas
@@ -18,8 +18,12 @@ import { AuthGuard } from './shared/_guards/auth.guard';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SharedModule } from './shared/shared.module';
 import { NgbdSortableHeader } from './config/sortable.directive';
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import localeEs from '@angular/common/locales/es-MX';
+import localeEsExtra from '@angular/common/locales/extra/es-MX';
+import { registerLocaleData } from '@angular/common';
 
+
+registerLocaleData(localeEs, 'es-MX', localeEsExtra);
 
 export function getToken() {
   return localStorage.getItem('token');
@@ -27,19 +31,20 @@ export function getToken() {
 
 @NgModule({
   declarations: [
+    // Todo lo que sirve dentro de este scope / modulo
     AppComponent,
     AuthComponent,
     PagesComponent,
-    NgbdSortableHeader
+    NgbdSortableHeader,
   ],
   imports: [
+    // Cuando incluyan la palabra modulo eva en esta area
     BrowserModule,
     BrowserAnimationsModule,
     FormsModule,
     ReactiveFormsModule,
     HttpClientModule,
     APP_ROUTES,
-    NgbModule,
     CoreModule,
     SharedModule,
     JwtModule.forRoot({
@@ -49,16 +54,17 @@ export function getToken() {
         blacklistedRoutes: ['localhost:5000/api/auth']
       }
     }),
-    // este seria ShareModule
   ],
   exports: [
-    BrowserAnimationsModule,
-    NgbdSortableHeader
+    // Lo que puedo utilizar en otros componentes.
+    BrowserAnimationsModule
   ],
   providers: [
+    { provide: LOCALE_ID, useValue: 'es-MX' },
     ErrorInterceptorProvider,
     AuthGuard
   ],
   bootstrap: [AppComponent]
 })
+
 export class AppModule { }
