@@ -1,5 +1,4 @@
 import { SortDirection } from './sortable.directive';
-import { PipeTransform } from '@angular/core';
 
 export interface ResBusqueda {
     obj: any[];
@@ -28,23 +27,23 @@ export class PaginacionRes<T> {
     paginacion: Paginacion;
 }
 
-function compare(v1, v2) {
-    return v1 < v2 ? -1 : v1 > v2 ? 1 : 0;
-  }
+// function compare(v1, v2) {
+//     return v1 < v2 ? -1 : v1 > v2 ? 1 : 0;
+//   }
 
-  function sort(obj: any[], column: string, direction: string): any[] {
-    if (direction === '') {
-      return obj;
-    } else {
-      return [...obj].sort((a, b) => {
-        const res = compare(a[column], b[column]);
-        return direction === 'asc' ? res : -res;
-      });
-    }
-  }
+  // function sort(obj: any[], column: string, direction: string): any[] {
+  //   if (direction === '') {
+  //     return obj;
+  //   } else {
+  //     return [...obj].sort((a, b) => {
+  //       const res = compare(a[column], b[column]);
+  //       return direction === 'asc' ? res : -res;
+  //     });
+  //   }
+  // }
 
-  function matches(obj: any, term: string, pipe: PipeTransform) {
-    return obj.name.toLowerCase().includes(term)
-      || pipe.transform(obj.area).includes(term)
-      || pipe.transform(obj.population).includes(term);
-  }
+  // function matches(obj: any, term: string, pipe: PipeTransform) {
+  //   return obj.name.toLowerCase().includes(term)
+  //     || pipe.transform(obj.area).includes(term)
+  //     || pipe.transform(obj.population).includes(term);
+  // }Í
