@@ -1,42 +1,42 @@
-import { Injectable } from '@angular/core';
+// import { Injectable } from '@angular/core';
 
-@Injectable()
-export class SidebarService {
+// @Injectable()
+// export class SidebarService {
 
 
 
-  menu: any = [
-    {
-      titulo: 'Principal',
-      icono: 'mdi mdi-gauge',
-      submenu: [
-        { titulo: 'Dashboard', url: '/dashboard' },
-        { titulo : 'ProgressBar', url: '/progress' },
-        { titulo: 'Gráficas', url: '/graficas1' },
-        { titulo: 'Promesas', url: '/promesas' },
-        { titulo: 'RxJs', url: '/rxjs' }
-      ]
-    },
-    {
-      titulo: 'Mantenimientos',
-      icono: 'mdi mdi-folder-lock-open',
-      submenu: [
-        { titulo: 'Usuarios', url: '/users' },
-        { titulo: 'Hospitales', url: '/perfil' },
-        { titulo: 'Médicos', url: '/medicos' }
-      ]
-    }
-  ];
+//   menu: any = [
+//     {
+//       titulo: 'Principal',
+//       icono: 'mdi mdi-gauge',
+//       submenu: [
+//         { titulo: 'Dashboard', url: '/dashboard' },
+//         { titulo : 'ProgressBar', url: '/progress' },
+//         { titulo: 'Gráficas', url: '/graficas1' },
+//         { titulo: 'Promesas', url: '/promesas' },
+//         { titulo: 'RxJs', url: '/rxjs' }
+//       ]
+//     },
+//     {
+//       titulo: 'Mantenimientos',
+//       icono: 'mdi mdi-folder-lock-open',
+//       submenu: [
+//         { titulo: 'Usuarios', url: '/users' },
+//         { titulo: 'Hospitales', url: '/perfil' },
+//         { titulo: 'Médicos', url: '/medicos' }
+//       ]
+//     }
+//   ];
 
-  constructor(
-    // public _usuarioService: UsuarioService
+//   constructor(
+//     // public _usuarioService: UsuarioService
 
-  ) {
-    this.cargarMenu();
-   }
+//   ) {
+//     this.cargarMenu();
+//    }
 
-  cargarMenu() {
-    this.menu = this.menu;
-  }
+//   cargarMenu() {
+//     this.menu = this.menu;
+//   }
 
-}
+// }
